fixed gallery image load, fixed local pathing

+++ utilities.py
@@ -6,7 +6,7 @@ from io import BytesIO
 import requests
 import os
 from PIL import Image
-from pathlib import Path
+from github import Github
 # Utility Functions to be called from all modules
 
 # Function to return a list of keys of a nested dictionary using it's key value (item or creature)
@@ -37,14 +37,15 @@ def directory_contents(directory_path):
 
 # Delete a list of file 
 def delete_files(file_paths):
+    if file_paths:
      
-    for file_path in file_paths:
-        try:
-            os.remove(f"./image_temp/{file_path}")
-            print(f"Remove : ./image_temp/{file_path}")
-        except OSError as e:
-            print(f"Error: {file_path} : {e.strerror}")
-    file_paths.clear()
+        for file_path in file_paths:
+            try:
+                os.remove(f"./image_temp/{file_path}")
+                print(f"Remove : ./image_temp/{file_path}")
+            except OSError as e:
+                print(f"Error: {file_path} : {e.strerror}")
+        file_paths.clear()
     
 
 def open_image_from_url(image_url):
@@ -59,4 +60,17 @@ def receive_upload(image_file):
 
     print(image)
     return image
+
+def index_image_paths(repo_name,directory_path):
+    g = Github()  # No token needed for public repos
+    repo = g.get_repo(repo_name)
+    contents = repo.get_contents(directory_path)
+
+    files = []
+    for content_file in contents:
+        if content_file.type == "file":
+            media_url = content_file.download_url.replace("raw.githubusercontent.com", "media.githubusercontent.com/media")
+            files.append(media_url)  # Or content_file.path for just the path
+    
+    return files
